services/notification_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
PyMySQL==1.1.0
cryptography==41.0.4
python-dotenv==1.0.0
nats-py==2.7.2
